Remove re-export of configuration
